stm | menu - moved menu functinos to seperate file `menu.c`, merged master
