+++ DockerControlSTM/Src/main.c
@@ -48,6 +48,7 @@
 #include "usbd_cdc_if.h"
 
 #include "hd44780.h"
+#include "menu/menu.h"
 /* USER CODE END Includes */
 
 /* Private typedef -----------------------------------------------------------*/
@@ -78,13 +79,14 @@ volatile uint8_t positions;
 char name[48];
 volatile int name_pos = 0;
 uint8_t character[1] = {0};
-uint8_t shift = 0;
+bool shift = 0;
 uint8_t back = 0;
 const char * menu_debug = "";
 
-//const char * menu_first_line = "";
+//menu
+esp_param = PARAM_IP;
+int current_menu = MENU_START;
 
-//const char * menu_second_line = "";
 
 char menu_first_line[48];
 char menu_second_line[48];
@@ -104,10 +106,7 @@ extern char wifi_password;
 extern uint8_t packet_received;
 extern uint8_t received_packet_header[10];
 extern uint8_t received_packet_body[4096];
-char temp1[50];
-char temp2[50];
-char temp3[50];
-char temp4[50];
+
 
 extern uint8_t packet_header[10];
 extern uint8_t packet_body[4096];
@@ -131,22 +130,6 @@ const uint16_t uart_size = 5;
 uint8_t usb_data[40];
 uint8_t usb_received = 0;
 
-typedef enum{
-    MENU_START,
-    MENU_DEFAULT,
-    MENU_CUSTOM,
-    MENU_CUSTOM_SERVER,
-    MENU_CUSTOM_WIFI
-} menu;
-
-typedef enum{
-    PARAM_IP,
-    PARAM_PORT,
-    PARAM_NAME,
-    PARAM_PASSWORD
-} esp_params;
-esp_param = PARAM_IP;
-menu current_menu = MENU_START;
 /* USER CODE END PV */
 
 /* Private function prototypes -----------------------------------------------*/
@@ -157,209 +140,6 @@ static void MX_TIM1_Init(void);
 static void MX_TIM2_Init(void);
 static void MX_TIM4_Init(void);
 /* USER CODE BEGIN PFP */
-
-int button_enter(){
-  if(enter_pressed){
-    enter_pressed = 0;
-    return 1;
-  }
-  else{
-    return 0;
-  }
-}
-
-int button_confirm(){
-  if(confirm_pressed){
-    confirm_pressed = 0;
-    return 1;
-  }
-  else{
-    return 0;
-  }
-}
-
-void temp_confirm() {}
-
-menu show_menu(current){
-  switch (current_menu) {
-    case MENU_START:
-      switch (positions % 2) {
-        case 0:
-            strcpy(menu_first_line, ">Default");
-            strcpy(menu_second_line, " Custom");
-//          menu_first_line = ">Default";
-//          menu_second_line = " Custom";
-          if (button_enter()) {
-            current_menu = MENU_DEFAULT;
-          }
-        break;
-        case 1:
-            strcpy(menu_first_line, " Default");
-            strcpy(menu_second_line, ">Custom");
-//          menu_first_line = " Default";
-//          menu_second_line = ">Custom";
-          if (button_enter()) {
-            current_menu = MENU_CUSTOM;
-          }
-        break;
-        default:break;
-      }
-    break;
-    case MENU_DEFAULT:
-//      menu_first_line = "Connecting using defaults";
-    break;
-    case MENU_CUSTOM:
-      switch (positions % 4) {
-        case 0:
-            strcpy(menu_first_line, ">Server IP: ");
-            strcat(menu_first_line, server_ip);
-            strcpy(menu_second_line, " Server Port: ");
-              strcat(menu_second_line, server_port);
-//          menu_first_line = ">Server IP";
-//          menu_second_line = " Server port";
-          if (button_enter()) {
-            esp_param = PARAM_IP;
-            current_menu = MENU_CUSTOM_SERVER;
-
-          }
-          else if(button_confirm()){
-            temp_confirm();
-            current_menu = MENU_START;
-          }
-          break;
-        case 1:
-          strcpy(menu_first_line, " Server IP: ");
-              strcat(menu_first_line, server_ip);
-              strcpy(menu_second_line, ">Server Port: ");
-              strcat(menu_second_line, server_port);
-//          menu_first_line = ">Server port";
-//          menu_second_line = " WiFi Name";
-          if (button_enter()) {
-            esp_param = PARAM_PORT;
-            current_menu = MENU_CUSTOM_SERVER;
-          }
-          else if(button_confirm()){
-            temp_confirm();
-            current_menu = MENU_START;
-          }
-          break;
-        case 2:
-            strcpy(menu_first_line, ">WiFi Name: ");
-            strcat(menu_first_line, wifi_name);
-            strcpy(menu_second_line, " WiFi Password: ");
-            strcat(menu_second_line, wifi_password);
-//          menu_first_line = ">WiFi Name";
-//          menu_second_line = " WiFi Password";
-          if (button_enter()) {
-            esp_param = PARAM_NAME;
-            current_menu = MENU_CUSTOM_WIFI;
-          }
-          else if(button_confirm()){
-            temp_confirm();
-            current_menu = MENU_START;
-          }
-          break;
-        case 3:
-            strcpy(menu_first_line, " WiFi Name: ");
-              strcat(menu_first_line, wifi_name);
-            strcpy(menu_second_line, ">WiFi Password: ");
-              strcat(menu_second_line, wifi_password);
-//          menu_first_line = " WiFi Name";
-//          menu_second_line = ">WiFi Password";
-          if (button_enter()) {
-            esp_param = PARAM_PASSWORD;
-            current_menu = MENU_CUSTOM_WIFI;
-          }
-          else if(button_confirm()){
-            temp_confirm();
-            current_menu = MENU_START;
-          }
-          break;
-        default:break;
-      }
-    break;
-    case MENU_CUSTOM_SERVER:
-      character[0] = (uint8_t) (positions%10 + 48);
-//          character[0] = (uint8_t) (positions + 65 + (shift * 32));
-      strcpy(menu_first_line, "Number: ");
-//          strcat(menu_first_line, (const char *) (character[0]-2));
-//          strcat(menu_first_line, " ");
-//          strcat(menu_first_line, (const char *) (character[0]-1));
-          strcat(menu_first_line, ">");
-          strcat(menu_first_line, (const char *) character);
-          strcat(menu_first_line, "<");
-//          strcat(menu_first_line, (const char *) (character[0]+1));
-//          strcat(menu_first_line, " ");
-//          strcat(menu_first_line, (const char *) (character[0]+2));
-          strcat(menu_first_line, " ");
-      strcpy(menu_second_line, name);
-      if(button_enter()){
-        name[name_pos] = character[0];
-        name_pos++;
-      }
-      if(shift == 1){
-        shift = 0;
-        name[name_pos] = '.';
-        name_pos++;
-      }
-      if(back_pressed == 1){
-        back_pressed = 0;
-        name_pos--;
-        name[name_pos] = ' ';
-      }
-      if(button_confirm()){
-        switch(esp_param){
-            case PARAM_IP:
-              strcpy(temp1,name);
-              server_ip = temp1;
-              break;
-          case PARAM_PORT:
-                strcpy(temp2, name);
-                server_port = temp2;
-                break;
-        }
-        current_menu = MENU_CUSTOM;
-      }
-      shift = 0;
-    break;
-    case MENU_CUSTOM_WIFI:
-      character[0] = (uint8_t) ((positions + 65 + shift*32));
-          strcpy(menu_first_line, "Letter: ");
-          strcat(menu_first_line, ">");
-          strcat(menu_first_line, (const char *) character);
-          strcat(menu_first_line, "<");
-          strcpy(menu_second_line, name);
-          if(button_enter()){
-            name[name_pos] = character[0];
-            name_pos++;
-          }
-          if(shift == 1){
-            shift = 0;
-            name[name_pos] = '.';
-            name_pos++;
-          }
-          if(back_pressed == 1){
-            back_pressed = 0;
-            name_pos--;
-            name[name_pos] = ' ';
-          }
-          if(button_confirm()){
-            switch(esp_param){
-              case PARAM_NAME:
-                    strcpy(temp3, name);
-                    wifi_name = temp3;
-                    break;
-              case PARAM_PASSWORD:
-                    strcpy(temp4, name);
-                    wifi_password = temp4;
-                    break;
-            }
-            current_menu = MENU_CUSTOM;
-          }
-          shift = 0;
-          break;
-  }
-}
 
 void start_dc() {
     HAL_UART_Receive_IT(&huart3, uart_receive, uart_size);
@@ -415,7 +195,7 @@ void HAL_TIM_PeriodElapsedCallback(TIM_HandleTypeDef *htim){
     HAL_TIM_Base_Stop(&htim2);
   }
   if(htim->Instance == TIM4){
-    show_menu(current_menu);
+    show_menu();
 //    current_menu = show_menu(current_menu);
     pulse_count = TIM1->CNT;
     positions = pulse_count/4;
